testutil: move to a submodule of spread
